Don't run linter if uncrustify version is too crusty.

+++ test/lint/run-linter.bash
@@ -36,13 +36,27 @@ for c_file in src/*.c; do
     rm $c_file.uncrustify
 done
 
+# version check
+temp_file=$(mktemp)
+uncrustify --version > ${temp_file}
+sed -i -e 's/Uncrustify-//g' ${temp_file}
+sed -i -e 's/_[a-zA-Z]//g' ${temp_file}
+version=$(cat ${temp_file})
+rm -f ${temp_file}
+
+if [[ $version < 0.65 ]]; then
+    echo -e "\e[31mUncrustify cannot test headers!\e[0m"
+    echo -e "\e[32mPretending things are fine...\e[0m"
+    exit 0
+fi
+
 # Lint the headers
 for c_file in include/*/*.h; do
     uncrustify -c $config $c_file
     dif=$(diff -ur $c_file $c_file.uncrustify)
 
     if [[ ! -z $dif ]]; then
-        echo -e "\[31m$c_file does not conform to Kernel style!"
+        echo -e "\e[31m$c_file does not conform to Kernel style!"
         echo "Patch will be found in $c_file.patch."
         diff -ur $c_file $c_file.uncrustify > $c_file.patch
         rm $c_file.uncrustify
